+++ misoclib/tools/litescope/example_designs/test/test_regs.py
@@ -4,6 +4,6 @@ def main(wb):
 	###
 	print("sysid     : 0x{:04x}".format(regs.identifier_sysid.read()))
 	print("revision  : 0x{:04x}".format(regs.identifier_revision.read()))
-	print("frequency : 0x{:04x}MHz".format(regs.identifier_frequency.read()/1000000))
+	print("frequency : 0x{:04x}MHz".format(int(regs.identifier_frequency.read()/1000000)))
 	###
 	wb.close()
